vendor: Fix PRODUCT_PROPERTY_OVERRIDES
This file gets included in config/common.mk, where we have other
property values set (such as ro.build.selinux for the status in the
About phone section). The problem with ":=" is it will override the
previous value with the new one, where as "+=" will add the supplied
value on to the existing one. With this change, all property values
are properly added!

Change-Id: Ibc32d1b8155bcd391dd95ab8b71d984c0dd54a99
Fixes: 56400a32 ("Set new pixel sounds as defaults")
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>